Rename modules component_type to schema_type and component to schema
