+++ utoipa-gen/src/schema/into_params.rs
@@ -7,7 +7,7 @@ use syn::{
 };
 
 use crate::{
-    component_type::{SchemaFormat, SchemaType},
+    schema_type::{SchemaFormat, SchemaType},
     doc_comment::CommentAttributes,
     parse_utils,
     path::parameter::{ParameterExt, ParameterIn, ParameterStyle},
